refactor: extract parsers into own directory
