+++ lib/constants.ts
@@ -1,0 +1,8 @@
+export const categoryNames: Record<string, string> = {
+  api: 'API',
+  'api/structures': 'API Structures',
+  development: 'Development',
+  tutorial: 'Guides',
+}
+
+export const IGNORE_PATTERN = '<!-- i18n-ignore -->'
